Refactor stats exporters into ext packages
