+++ contrib/opencensus-ext-prometheus/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## Unreleased
+
+- Initial version.
