+++ contrib/opencensus-ext-prometheus/setup.cfg
@@ -1,0 +1,2 @@
+[bdist_wheel]
+universal = 1
