+++ nox.py
@@ -31,6 +31,7 @@ def _install_dev_packages(session):
     session.install('-e', 'contrib/opencensus-ext-mysql')
     session.install('-e', 'contrib/opencensus-ext-ocagent')
     session.install('-e', 'contrib/opencensus-ext-postgresql')
+    session.install('-e', 'contrib/opencensus-ext-prometheus')
     session.install('-e', 'contrib/opencensus-ext-pymongo')
     session.install('-e', 'contrib/opencensus-ext-pymysql')
     session.install('-e', 'contrib/opencensus-ext-pyramid')
