+++ requirements-test.txt
@@ -6,5 +6,4 @@ pytest-cov==2.5.1
 retrying==1.3.3
 WebOb==1.7.3
 thrift==0.10.0
-prometheus_client==0.5.0
 unittest2==1.1.0
